delete

+++ shopping/src/main/java/com/a/shopping/repository/ProductImageRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
-    List<ProductImage> findByProductId(Integer productId);
+    List<ProductImage> findByProductId(long productId);
 }
